requirements.txt
Package              Version
-------------------- ---------
Authlib              1.2.0
cachetools           5.3.0
certifi              2022.12.7
cffi                 1.15.1
charset-normalizer   3.1.0
click                8.1.3
cryptography         40.0.1
Flask                2.2.3
google-auth          2.16.2
google-auth-oauthlib 1.0.0
gspread              5.7.2
idna                 3.4
itsdangerous         2.1.2
Jinja2               3.1.2
Markdown             3.4.3
MarkupSafe           2.1.2
oauthlib             3.2.2
pip                  23.0.1
pyasn1               0.4.8
pyasn1-modules       0.2.8
pycparser            2.21
python-dotenv        1.0.0
requests             2.28.2
requests-oauthlib    1.3.1
rsa                  4.9
setuptools           63.2.0
six                  1.16.0
urllib3              1.26.15
Werkzeug             2.2.3
